Simulator interface almost done missing some points
Created a new module for the lemonator that houses a Lemonator class
Recreated the c++ interfaces in python with some adjustments so the
simulator plays nicely.

Changed the effectors, sensors to only take a vessel in the update.
so the constructor isn't vessel bound but update can still use it

Updated the gui to work with the new interface and also be conform the c++
now

Updated simulator, vessel, plant to work with the lemonator object

TODO: Fix measurements and actually simulate something nice
TODO: valves
TODO: tap
TODO: test it with the pyd